app/requirements.txt
requests
boto3
warrant
python-dotenv
